add userRequestDto and userResponseDto, modify application.properties, add ignore files
