+++ .gitignore
@@ -1,4 +1,5 @@
 target/
+data/
 !.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**/target/
 !**/src/test/**/target/
