+++ pom.xml
@@ -82,6 +82,14 @@
       <version>1.4.197</version>
       <scope>runtime</scope>
     </dependency>
+
+    <!-- Driver MySQL -->
+    <dependency>
+      <groupId>mysql</groupId>
+      <artifactId>mysql-connector-java</artifactId>
+    </dependency>
+
+    <!-- validator de entidades-->
     <dependency>
       <groupId>org.hibernate.validator</groupId>
       <artifactId>hibernate-validator</artifactId>
