+++ src/main/java/org/drg/App.java
@@ -1,9 +1,6 @@
 package org.drg;
 
-import org.drg.model.entity.UserEntity;
-import org.drg.repository.UserRepository;
 import org.modelmapper.ModelMapper;
-import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
@@ -18,17 +15,6 @@ public class App
     public static void main( String[] args )
     {
         SpringApplication.run(App.class, args);
-    }
-
-    @Bean
-    public CommandLineRunner run(UserRepository userRepository) throws Exception {
-        return (String[] args) -> {
-            UserEntity user1 = new UserEntity("Bob", "bob@domain.com");
-            UserEntity user2 = new UserEntity("Jenny", "jenny@domain.com");
-            userRepository.save(user1);
-            userRepository.save(user2);
-            userRepository.findAll().forEach(System.out::println);
-        };
     }
 
     @Bean
