--- src/main/java/org/drg/controller/IUserController.java
@@ -1,35 +1,0 @@
-package org.drg.controller;
-
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.enums.ParameterIn;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import org.drg.model.dto.request.UserRequestDto;
-import org.drg.model.dto.response.UserResponseDto;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-
-import javax.validation.Valid;
-
-@Validated
-public interface IUserController {
-
-    @Operation(summary = "", description = "Creates an user and returns the user ID.", tags={ "users" })
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "201", description = "Created user", content = @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponseDto.class))),
-            @ApiResponse(responseCode = "400", description = "Bad request.", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))),
-            @ApiResponse(responseCode = "409", description = "Username already exists.", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))),
-            @ApiResponse(responseCode = "500", description = "Internal error.", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))) })
-    @RequestMapping(value = "/users",
-            produces = { "application/json" },
-            consumes = { "application/json" },
-            method = RequestMethod.POST)
-    ResponseEntity<UserResponseDto> createUser(@Parameter(in = ParameterIn.DEFAULT, description = "A JSON object that contains the username (email) and password.", required=true, schema=@Schema()) @Valid @RequestBody UserRequestDto body);
-
-}
