--- src/main/java/org/drg/controller/impl/UserController.java
@@ -1,35 +1,0 @@
-package org.drg.controller.impl;
-
-import org.drg.model.entity.UserEntity;
-import org.drg.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-import javax.validation.Valid;
-import java.util.List;
-
-@RestController
-public class UserController {
-
-    @Autowired
-    private UserService userService;
-
-    @GetMapping("/users")
-    public List<UserEntity> getUsers() {
-        return (List<UserEntity>) userService.findAll();
-    }
-
-    @PostMapping("/users")
-    ResponseEntity<String> addUser(@Valid @RequestBody UserEntity user) {
-        userService.saveUser(user);
-        return ResponseEntity.ok("User is valid");
-    }
-
-
-
-}
-
