+++ src/main/java/org/drg/dtos/ErrorsDto.java
@@ -1,6 +1,7 @@
-package org.drg.handlerError.dto;
+package org.drg.dtos;
 
 import lombok.Data;
+import org.drg.dtos.ErrorDto;
 
 import java.util.List;
