+++ src/main/java/org/drg/dtos/requests/UserRequestDto.java
@@ -1,13 +1,18 @@
-package org.drg.model.dto.response;
+package org.drg.dtos.requests;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.springframework.validation.annotation.Validated;
 
 @Validated
 @javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-11-02T19:44:59.902Z[GMT]")
 
-
-public class UserResponseDto {
+@Getter
+@Setter
+@NoArgsConstructor
+public class UserRequestDto {
 
     private Long id;
     @JsonProperty("username")
@@ -31,10 +36,9 @@ public class UserResponseDto {
     @JsonProperty("door")
     private Object door;
 
-    public UserResponseDto username(String username) {
+    public UserRequestDto username(String username) {
         this.username = username;
         return this;
     }
-
 
 }
