--- src/main/java/org/drg/exception/MyResourceNotFoundException.java
@@ -1,25 +1,0 @@
-package org.drg.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public final class MyResourceNotFoundException extends RuntimeException {
-
-    public MyResourceNotFoundException() {
-        super();
-    }
-
-    public MyResourceNotFoundException(final String message, final Throwable cause) {
-        super(message, cause);
-    }
-
-    public MyResourceNotFoundException(final String message) {
-        super(message);
-    }
-
-    public MyResourceNotFoundException(final Throwable cause) {
-        super(cause);
-    }
-
-}
