+++ src/main/java/org/drg/repositories/UserRepository.java
@@ -1,0 +1,8 @@
+package org.drg.repositories;
+
+import org.drg.models.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<UserEntity, Long> {}
