--- src/main/java/org/drg/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package org.drg.repository;
-
-import org.drg.model.entity.UserEntity;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends CrudRepository<UserEntity, Long> {}
