--- src/main/java/org/drg/service/impl/UserServiceImpl.java
@@ -1,26 +1,0 @@
-package org.drg.service.impl;
-
-import org.drg.model.entity.UserEntity;
-import org.drg.repository.UserRepository;
-import org.drg.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-@Service
-public class UserServiceImpl implements UserService {
-
-    @Autowired
-    private UserRepository userRepository;
-
-
-    @Override
-    public List<UserEntity> findAll() {
-        return (List<UserEntity>) userRepository.findAll();
-    }
-
-    @Override
-    public void saveUser(UserEntity user) {
-        userRepository.save(user);
-    }
-}
