+++ src/main/resources/application-dev.properties
@@ -1,0 +1,25 @@
+# spring traces
+logging.level.org.springframework.web=DEBUG
+
+# open-api
+springdoc.api-docs.path=/api-docs
+springdoc.swagger-ui.path=/springbooks-manager-api.html
+
+# jpa
+spring.jpa.hibernate.ddl-auto=create-drop
+spring.datasource.initialization-mode=always
+spring.jpa.properties.hibernate.format_sql=true
+logging.level.org.hibernate.SQL=DEBUG
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
+
+# database connection
+spring.datasource.url=jdbc:mysql://localhost/BD-COURT
+spring.datasource.username=root
+spring.datasource.password=pass
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+
+# https
+server.port = 8081
+#server.ssl.key-store = classpath:keystore.jks
+#server.ssl.key-store-password = password
+#server.ssl.key-password = secret
