+++ src/main/resources/application-local.properties
@@ -1,0 +1,32 @@
+# spring traces
+logging.level.org.springframework.web=DEBUG
+
+# open-api
+springdoc.api-docs.path=/api-docs
+springdoc.swagger-ui.path=/springbooks-manager-api.html
+
+# jpa
+spring.jpa.hibernate.ddl-auto=create-drop
+spring.datasource.initialization-mode=always
+spring.jpa.properties.hibernate.format_sql=true
+logging.level.org.hibernate.SQL=DEBUG
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
+spring.jpa.database-platform= org.hibernate.dialect.H2Dialect
+spring.jpa.defer-datasource-initialization= true
+
+
+# database connection
+spring.datasource.url=jdbc:h2:file:./data/demo
+spring.datasource.username=sa
+spring.datasource.password=password
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.h2.console.enabled= true
+
+# https
+server.port = 8081
+#server.ssl.key-store = classpath:keystore.jks
+#server.ssl.key-store-password = password
+#server.ssl.key-password = secret
+
+
+
