--- src/main/resources/application-local.yml
@@ -1,15 +1,0 @@
-server:
-  port: 8081
-
-spring:
-    datasource:
-      url: jdbc:h2:file:./data/demo
-      driverClassName: org.h2.Driver
-      username: sa
-      password: password
-      h2:
-        console:
-          enabled:true:
-      jpa:
-        database-platform: org.hibernate.dialect.H2Dialect
-        defer-datasource-initialization: true
