+++ src/main/resources/data.sql
@@ -1,0 +1,5 @@
+
+
+-- saving users
+INSERT INTO t_user (id, mail, name) VALUES(1, 'user1@email.es', 'user1');
+INSERT INTO t_user (id, mail, name) VALUES(2, 'user2@email.es', 'user2');
